requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pyyaml>=5.4.1
aiofiles>=0.7.0
jinja2>=3.0.0
python-multipart>=0.0.5
websockets>=10.0
psutil>=5.8.0